refactor

+++ LeetCode/162_寻找峰值.cpp
@@ -17,6 +17,6 @@ public:
                 right = middle;
             }
         }
-        return left;
+        return static_cast<int>(left);
     }
 };
