+++ LeetCode/268_丢失的数字.cpp
@@ -6,7 +6,7 @@ class Solution {
 public:
     int missingNumber(vector<int>& nums) {
         int sum = accumulate(nums.begin(), nums.end(), 0);
-        int total_sum = (nums.size() * (nums.size() + 1)) / 2;
+        int total_sum = static_cast<int>(nums.size() * (nums.size() + 1)) / 2;
         return total_sum - sum;
     }
 };
