+++ LeetCode/407_接雨水_II.cpp
@@ -16,18 +16,18 @@ class Solution {
     };
 public:
     int trapRainWater(vector<vector<int>>& heightMap) {
-        int m = heightMap.size();
-        int n = heightMap[0].size();
+        size_t m = heightMap.size();
+        size_t n = heightMap[0].size();
         vector<vector<bool>> visit(m, vector<bool>(n));
         priority_queue<Node,vector<Node>, NodeCmp> q;
         for (int i = 0; i < m; i++) {
             q.push({ i, 0, heightMap[i][0] });
-            q.push({ i, n - 1, heightMap[i][n - 1] });
+            q.push({ i, static_cast<int>(n - 1), heightMap[i][n - 1] });
             visit[i][0] = visit[i][n - 1] = true;
         }
         for (int j = 0; j < n; j++) {
             q.push({ 0, j, heightMap[0][j] });
-            q.push({ m - 1,j, heightMap[m - 1][j] });
+            q.push({ static_cast<int>(m - 1),j, heightMap[m - 1][j] });
             visit[0][j] = visit[m - 1][j] = true;
         }
         int dirs[4][2] = { {1,0},{-1,0},{0,1},{0,-1} };
