+++ LeetCode/434_字符串中的单词数.cpp
@@ -11,6 +11,6 @@ public:
         while (in >> word) {
             ++count;
         }
-        return count;
+        return static_cast<int>(count);
     }
 };
