+++ LeetCode/488_祖玛游戏.cpp
@@ -80,7 +80,7 @@ public:
                 }
             }
         }
-        return min_count;
+        return static_cast<int>(min_count);
     }
 };
