+++ LeetCode/575_分糖果.cpp
@@ -6,6 +6,6 @@ class Solution {
 public:
     int distributeCandies(vector<int>& candyType) {
         unordered_set<int> candy(candyType.begin(), candyType.end());
-        return min(candy.size(), candyType.size() / 2);
+        return static_cast<int>(min(candy.size(), candyType.size() / 2));
     }
 };
