+++ LeetCode/58_最后一个单词的长度.cpp
@@ -8,6 +8,6 @@ public:
         istringstream is(s);
         string word;
         while (is >> word);
-        return word.length();
+        return static_cast<int>(word.length());
     }
 };
