+++ LeetCode/638_大礼包.cpp
@@ -7,13 +7,13 @@ public:
     map<vector<int>, int> memo;
 
     int shoppingOffers(vector<int>& price, vector<vector<int>>& special, vector<int>& needs) {
-        int n = price.size();
+        size_t n = price.size();
 
         // 过滤不需要计算的大礼包，只保留需要计算的大礼包
         vector<vector<int>> filterSpecial;
         for (auto& sp : special) {
             int totalCount = 0, totalPrice = 0;
-            for (int i = 0; i < n; ++i) {
+            for (size_t i = 0; i < n; ++i) {
                 totalCount += sp[i];
                 totalPrice += sp[i] * price[i];
             }
@@ -26,7 +26,7 @@ public:
     }
 
     // 记忆化搜索计算满足购物清单所需花费的最低价格
-    int dfs(vector<int> price, const vector<vector<int>>& special, vector<int> curNeeds, vector<vector<int>>& filterSpecial, int n) {
+    int dfs(vector<int> price, const vector<vector<int>>& special, vector<int> curNeeds, vector<vector<int>>& filterSpecial, size_t n) {
         if (!memo.count(curNeeds)) {
             int minPrice = 0;
             for (int i = 0; i < n; ++i) {
